+++ CHANGELOG.md
@@ -2,10 +2,32 @@
 
 All notable changes to `vCardIO` will be documented in this file.
 
+## 1.1.1 - 2024-11-03
+
+### Export methods
+
+On vCard object:
+
+```php
+$vCard->export('./file/path.vcf');
+
+```
+On vCards collection:
+
+```php
+$vCardCollection->export('./file/path.vcf', append:false);
+
+```
+**Full Changelog**: https://github.com/PierreLebedel/vCardIO/compare/1.1.0...1.1.1
+
 ## [1.1.0] - 2024-10-30
+
 ### Changed
+
 - Add vCard methods to get relevant data
 
 ## [1.0.0] - 2024-10-29
+
 ### Changed
+
 - Create project first release